backend/requirements.txt
Django>=4.0,<5.0
djangorestframework
environs
channels>=4.0.0
channels_redis>=4.0.0
daphne>=4.0.0
psycopg2
Pillow>=9.0.0
djangorestframework-simplejwt
django-cors-headers